feat(cairo_native): apply resource limits to compilation processes

+++ crates/native_blockifier/src/py_objects.rs
@@ -170,7 +170,7 @@ impl From<PyConcurrencyConfig> for ConcurrencyConfig {
 pub struct PySierraCompilationConfig {
     pub sierra_to_native_compiler_path: String,
     pub libcairo_native_runtime_path: String,
-    pub max_native_bytecode_size: usize,
+    pub max_native_bytecode_size: u64,
     pub max_cpu_time: u64,
     pub max_memory_usage: u64,
 }
